Update view and mainlander, added two gifs
Massive changes to both view and moonlander. Game is in better state than it was. In order to be working at its core, we need the methods move, thrust, and changeVelocity in the MoonLander class to be reconfigured to work properly. Just about everything else works, other than working out kinks for end of game. Just pull everything from all files to continue work, the only thing I did not touch was the Leaderboard class.